requirements.txt
pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu117
scipy
matplotlib
scikit-learn
einops
torch_scatter
torch_geometric
numpy